FewBodyDB.jl

This package is a database of calculation results for few-body systems. It was developed for testing solvers in JuliaFewBody projects.

Usage

The @get macro returns the bibliographic value corresponding to the given key.

julia> using FewBodyDB
julia> @get Bubin2005Jan, HD⁺, energy, (J=0, v=0)"-0.5978979685"
julia> println(@bib Bubin2005Jan)@article{Bubin2005Jan,
  title = {Charge asymmetry in HD+},
  volume = {122},
  ISSN = {1089-7690},
  url = {http://dx.doi.org/10.1063/1.1850905},
  DOI = {10.1063/1.1850905},
  number = {4},
  journal = {The Journal of Chemical Physics},
  publisher = {AIP Publishing},
  author = {Bubin,  Sergiy and Bednarz,  Eugeniusz and Adamowicz,  Ludwik},
  year = {2005},
  month = jan 
}

Data

command	value

@get Suzuki2003Jul, ∞Li, energy, ¹Sᵉ	"-7.279913"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 17)	"-0.5042821328684"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 10)	"-0.52923363556611"

@get Suzuki2003Jul, Ps⁻, energy, ¹Sᵉ	"-0.2620050702328"

@get Suzuki2003Jul, ttμ, energy, ¹Sᵉ	"-112.9730179"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 8)	"-0.53968792704413"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 3)	"-0.57305054655187"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 10)	"-0.52911665292566"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 17)	"-0.50445269886425"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 9)	"-0.53433701356127"

@get Suzuki2003Jul, tdμ, energy, ¹Dᵉ	"-103.408481"

@get Suzuki2003Jul, H⁻, energy, ³Pᵉ	"-0.1252865"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 6)	"-0.55149117284"

@get Suzuki2003Jul, ∞He, energy, ¹Sᵉ	"-2.903724376984"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 14)	"-0.51240752360873"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 11)	"-0.52440184594275"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 2)	"-0.58090370021837"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 20)	"-0.50029245420662"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 18)	"-0.502449537962"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 6)	"-0.55193594895682"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 19)	"-0.50119479911831"

@get Suzuki2003Jul, He, energy, ¹Fᵒ	"-2.031255144382"

@get Suzuki2003Jul, p̄He⁺, energy, L=31	"-3.50763486"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 20)	"-0.50026932312565"

@get Suzuki2003Jul, He, energy, ¹Sᵉ	"-2.903724372437"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 17)	"-0.50439557371086"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 8)	"-0.53942340521103"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 21)	"-0.4999027831355"

@get Suzuki2003Jul, tdμ, energy, ¹Pᵒ	"-108.179385"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 21)	"-0.499889109"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 5)	"-0.55810654817382"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 7)	"-0.54554530340943"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 20)	"-0.5002241584"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 4)	"-0.56561104207681"

@get Bubin2005Jan, HD⁺, energy, (J = 0, v = 1)	"-0.5891818291"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 22)	"-0.499864342"

@get Suzuki2003Jul, He, energy, ³Pᵒ	"-2.123843086498"

@get Suzuki2003Jul, ∞He, energy, ³Dᵉ	"-2.055338993337"

@get Suzuki2003Jul, ∞H⁻, energy, ¹Sᵉ	"-0.527751016523"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 7)	"-0.54568591529293"

@get Bubin2005Jan, HD⁺, energy, (J = 0, v = 2)	"-0.5809037001"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 10)	"-0.52888354391503"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 12)	"-0.52017114777622"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 5)	"-0.55841886325887"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 0)	"-0.59789796860903"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 4)	"-0.56544616627757"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 9)	"-0.53421216210065"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 22)	"-0.4998657785"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 14)	"-0.51257570550594"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 12)	"-0.52007010037534"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 2)	"-0.58035919519988"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 4)	"-0.56511744976374"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 1)	"-0.58861082938979"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 13)	"-0.5161258334232"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 7)	"-0.54526501568956"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 15)	"-0.50942885109613"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 3)	"-0.57253181032597"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 15)	"-0.50927803835031"

@get Suzuki2003Jul, He, energy, ³Fᵒ	"-2.031255168403"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 16)	"-0.5066971569796"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 21)	"-0.49991036147003"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 13)	"-0.51621870996171"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 6)	"-0.55178736790852"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 2)	"-0.58072182812093"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 11)	"-0.52451091017174"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 0)	"-0.59769812819221"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 1)	"-0.58899111199204"

@get Suzuki2003Jul, ttμ, energy, ¹Fᵒ	"-101.43"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 13)	"-0.51594088194548"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 1)	"-0.58918182955696"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 9)	"-0.53396333969469"

@get Suzuki2003Jul, ttμ, energy, ¹Dᵉ	"-105.98293"

@get Suzuki2003Jul, He, energy, ¹Gᵉ	"-2.020000710898"

@get Suzuki2003Jul, tdμ, energy, ¹Sᵉ	"-111.364511474"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 19)	"-0.501088766606"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 16)	"-0.5067638778173"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 19)	"-0.50115914795876"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 8)	"-0.53982064154582"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 14)	"-0.5126601922527"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 0)	"-0.59729964335178"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 5)	"-0.55857552082556"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 18)	"-0.50254238629066"

@get Suzuki2003Jul, He, energy, ³Gᵉ	"-2.020000710925"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 15)	"-0.50950465133511"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 12)	"-0.51986882186127"

@get Suzuki2003Jul, He, energy, ¹Dᵉ	"-2.055620732852"

@get Karr2006Apr, HD⁺, energy, (J = 0, v = 18)	"-0.50258923378925"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 11)	"-0.52418455220494"

@get Bubin2005Jan, HD⁺, energy, (J = 0, v = 0)	"-0.5978979685"

@get Suzuki2003Jul, ttμ, energy, ¹Pᵒ	"-110.2621165"

@get Suzuki2003Jul, He, energy, ³Sᵉ	"-2.175293782367"

@get Karr2006Apr, HD⁺, energy, (J = 1, v = 3)	"-0.57287727709421"

@get Karr2006Apr, HD⁺, energy, (J = 2, v = 16)	"-0.50656450765298"

Bibliography

The database references are here.

@article{Bubin2005Jan,
  title = {Charge asymmetry in HD+},
  volume = {122},
  ISSN = {1089-7690},
  url = {http://dx.doi.org/10.1063/1.1850905},
  DOI = {10.1063/1.1850905},
  number = {4},
  journal = {The Journal of Chemical Physics},
  publisher = {AIP Publishing},
  author = {Bubin,  Sergiy and Bednarz,  Eugeniusz and Adamowicz,  Ludwik},
  year = {2005},
  month = jan
}

@book{Suzuki2003Jul,
  title={Stochastic variational approach to quantum-mechanical few-body problems},
  volume = {54},
  ISSN = {0940-7677},
  url = {https://doi.org/10.1007/3-540-49541-X},
  DOI = {10.1007/3-540-49541-X},
  author={Suzuki, Yasuyuki and Varga, K{'a}lm{'a}n},
  year={2002},
  month = july,
  publisher={Springer},
}

@article{Karr2006Apr,
  title = {High accuracy results for the energy levels of the molecular ions H+2,  D+2and HD+,  up toJ= 2},
  volume = {39},
  ISSN = {1361-6455},
  url = {http://dx.doi.org/10.1088/0953-4075/39/8/024},
  DOI = {10.1088/0953-4075/39/8/024},
  number = {8},
  journal = {Journal of Physics B: Atomic,  Molecular and Optical Physics},
  publisher = {IOP Publishing},
  author = {Karr,  J Ph and Hilico,  L},
  year = {2006},
  month = apr,
  pages = {2095–2105}
}

Citation

Please use CITATION.bib if you need to cite this package.

@misc{FewBodyDB.jl,
	author  = {Shuhei Ohno, Martin Mikkelsen},
	title   = {FewBodyDB.jl},
	url     = {https://github.com/JuliaFewBody/FewBodyDB.jl},
	version = {v0.0.3},
	year    = {2025},
	month   = {9}
}

API reference

FewBodyDB.@bib

FewBodyDB.@get

FewBodyDB.@bib — Macro

Get BibTeX entry from FewBodyDB.DATA.

julia> println(@bib Bubin2005Jan)
@article{Bubin2005Jan,
  title = {Charge asymmetry in HD+},
  volume = {122},
  ISSN = {1089-7690},
  url = {http://dx.doi.org/10.1063/1.1850905},
  DOI = {10.1063/1.1850905},
  number = {4},
  journal = {The Journal of Chemical Physics},
  publisher = {AIP Publishing},
  author = {Bubin,  Sergiy and Bednarz,  Eugeniusz and Adamowicz,  Ludwik},
  year = {2005},
  month = jan 
}

FewBodyDB.@get — Macro

Get data from FewBodyDB.DATA.

julia> @get Bubin2005Jan, HD⁺, energy, (J=0, v=0)
"-0.5978979685"